Refactoring and now uses github api for downloading assets
