--- src/components/header/index.js
@@ -1,2 +1,0 @@
-import Header from "./header";
-export default Header;
