+++ src/model/status.js
@@ -1,0 +1,12 @@
+export const Status = {
+  DISCONNECTED: "disconnected",
+  CONNECTING: "connecting",
+  CONNECTED: "connected",
+  DOWNLOADING: "downloading",
+  EXTRACTING: "extracting",
+  FLASHING: "flashing",
+  DONE: "done",
+  ERROR: "error",
+};
+
+export default Status;
