--- src/panels/connection/index.js
@@ -1,2 +1,0 @@
-import Connection from "./connection";
-export default Connection;
