--- src/panels/done/index.js
@@ -1,2 +1,0 @@
-import Done from "./done";
-export default Done;
