--- src/panels/firmware/index.js
@@ -1,2 +1,0 @@
-import Firmware from "./firmware";
-export default Firmware;
