--- src/panels/progress/index.js
@@ -1,2 +1,0 @@
-import Progress from "./progress";
-export default Progress;
